+++ install.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright (C) 2016 Damon Lynch <damonlynch@gmail.com>
+# Copyright (C) 2016-2017 Damon Lynch <damonlynch@gmail.com>
 
 # This file is part of Rapid Photo Downloader.
 #
@@ -29,7 +29,7 @@ for Rapid Photo Downloader to run. Specifically, these packages are:
 1. Non-python programs, e.g. exiv2, ExifTool.
 2. Python packages that are unavailable on Python's PyPi service, namely
    python3 gobject introspection modules.
-3. Although PyQt 5.6 is available on PyPi, bundled with Qt 5.6, it's easier
+3. Although PyQt 5.6 and above is available on PyPi, bundled with Qt 5.6, it's easier
    to use the Linux distro's PyQt packages, particularly in the case of Ubuntu, whose
    custom scrollbar implementation does not work with stock Qt without a special environment
    variable being set that disables the custom scrollbars.
@@ -38,8 +38,8 @@ Once these dependencies are satisfied, Python's pip is used to install Rapid Pho
 itself, along with several Python packages from PyPi, found in requirements.txt.
 
 The secondary purpose of this install script is to give the option to the user of installing man
-pages in the system's standard  man page location, and for Debian/Ubuntu distros, to create a
-link in ~/bin to the rapid-photo-downloader executable.
+pages in the system's standard  man page location, and for Debian/Ubuntu/openSUSE distros,
+to create a link in ~/bin to the rapid-photo-downloader executable.
 """
 
 __author__ = 'Damon Lynch'
@@ -59,19 +59,6 @@ from enum import Enum
 
 
 try:
-    import pip
-except ImportError:
-    sys.stderr.write("To run this program, you must first install pip for Python 3\n")
-    sys.stderr.write("Install it using your Linux distribution's standard package manager -\n\n")
-    fedora_pip = 'Fedora:\nsudo dnf install python3 python3-wheel\n'
-    suse_pip = 'openSUSE:\nsudo zypper install python3-pip python3-setuptools python3-wheel\n'
-    debian_pip = 'Ubuntu/Debian/Mint:\nsudo apt-get install python3-pip python3-wheel\n'
-    arch_pip = 'Arch Linux:\nsudo pacman -S python-pip\n'
-    for distro in (fedora_pip, suse_pip, debian_pip, arch_pip):
-        sys.stderr.write('{}\n'.format(distro))
-    sys.exit(1)
-
-try:
     import apt
     have_apt = True
 except ImportError:
@@ -82,6 +69,9 @@ try:
     have_dnf = True
 except ImportError:
     have_dnf = False
+
+
+os_release = '/etc/os-release'
 
 
 class bcolors:
@@ -101,11 +91,22 @@ class Distro(Enum):
     fedora = 3
     neon = 4
     linuxmint = 5
-    unknown = 6
+    korora = 6
+    arch = 7
+    opensuse = 8
+    manjaro = 9
+    unknown = 20
 
 
 debian_like = (Distro.debian, Distro.ubuntu, Distro.neon, Distro.linuxmint)
+fedora_like = (Distro.fedora, Distro.korora)
+arch_like = (Distro.arch, Distro.manjaro)
 
+installer_cmds = {
+    Distro.fedora: 'dnf',
+    Distro.debian: 'apt-get',
+    Distro.opensuse: 'zypper',
+}
 
 def check_packages_on_other_systems() -> None:
     """
@@ -165,9 +166,11 @@ def get_distro_id(id_or_id_like: str) -> Distro:
         return Distro.unknown
 
 def get_distro() -> Distro:
-    if os.path.isfile('/etc/os-release'):
-        with open('/etc/os-release', 'r') as f:
+    if os.path.isfile(os_release):
+        with open(os_release, 'r') as f:
             for line in f:
+                if line.startswith('NAME=') and line.find('Korora') > 0:
+                    return Distro.korora
                 if line.startswith('ID='):
                     return get_distro_id(line[3:])
                 if line.startswith('ID_LIKE='):
@@ -175,67 +178,80 @@ def get_distro() -> Distro:
     return Distro.unknown
 
 def get_distro_version(distro: Distro) -> float:
+    remove_quotemark = False
     if distro == Distro.fedora:
         version_string = 'REDHAT_BUGZILLA_PRODUCT_VERSION='
-    elif distro in debian_like:
+    elif distro in debian_like or distro == Distro.opensuse:
         version_string = 'VERSION_ID="'
+        remove_quotemark = True
+    elif distro == Distro.korora:
+        version_string = 'VERSION_ID='
     else:
         return 0.0
 
-    with open('/etc/os-release', 'r') as f:
+    with open(os_release, 'r') as f:
         for line in f:
             if line.startswith(version_string):
                 try:
-                    if distro == Distro.fedora:
-                        v = line[len(version_string):]
-                    else:
+                    if remove_quotemark:
                         v = line[len(version_string):-2]
+                    else:
+                        v = line[len(version_string):]
                     return float(v)
                 except ValueError:
-                    sys.stderr.write("Unexpected format while parsing {} version".format(
+                    sys.stderr.write("Unexpected format while parsing {} version\n".format(
                         distro.name.capitalize()))
                     return 0.0
+    return 0.0
 
-def install_packages(command_line: str) -> None:
-    print("To continue, some packages required to run the application will be "
-          "installed.\n")
+
+def run_cmd(command_line: str, restart=False, exit_on_failure=True) -> None:
     print("The following command will be run:\n")
     print(command_line)
-    print("\nsudo may prompt you for the sudo password.\n")
-    answer = input('Do you agree to run this command? (if you do, type yes and hit '
-                   'enter): ')
-    if answer == 'yes':
-        args = shlex.split(command_line)
-        try:
-            subprocess.check_call(args)
-        except subprocess.CalledProcessError:
-            sys.stderr.write("Command failed: exiting\n")
-            sys.exit(1)
-    else:
-        print("Answer is not yes, exiting")
+    if command_line.startswith('sudo'):
+        print("\nsudo may prompt you for the sudo password.")
+    print()
+
+    args = shlex.split(command_line)
+    answer = input('Would you like to run the command now? (If you do, type yes and hit enter): ')
+
+    if answer != 'yes':
+        print('Answer is not yes, exiting.')
         sys.exit(0)
 
-def uninstall_packages(command_line: str) -> None:
-    print("\nThe following command will be run:\n")
-    print(command_line)
-    print("\nsudo may prompt you for the sudo password.\n")
-    answer = input('Do you agree to run this command? (if you do, type yes and hit enter): ')
-    if answer == 'yes':
-        args = shlex.split(command_line)
-        try:
-            subprocess.check_call(args)
-        except subprocess.CalledProcessError:
-            sys.stderr.write("Command failed\n")
     print()
+
+    try:
+        subprocess.check_call(args)
+    except subprocess.CalledProcessError:
+        sys.stderr.write("Command failed\n")
+        if exit_on_failure:
+            sys.stderr.write("Exiting\n")
+            sys.exit(1)
+    else:
+        if restart:
+            if len(sys.argv) == 2:
+                sys.stdout.flush()
+                sys.stderr.flush()
+                # restart the script
+                os.execl(sys.executable, sys.executable, *sys.argv)
+            else:
+                print("Rerun this script, passing the path to the tarfile\n")
+                sys.exit(0)
+
+
 
 def check_package_import_requirements(distro: Distro, version: float) -> None:
     if distro in debian_like:
         if not have_apt:
-            sys.stderr.write('To continue, the package python3-apt must be installed.')
-            sys.exit(1)
+            print('To continue, the package python3-apt must be installed.\n')
+            cmd = shutil.which('apt-get')
+            command_line = 'sudo {} install python3-apt'.format(cmd)
+            run_cmd(command_line, restart=True)
+
         cache = apt.Cache()
         missing_packages = []
-        packages = 'libimage-exiftool-perl python3-pyqt5 python3-setuptools python3-dev ' \
+        packages = 'libimage-exiftool-perl python3-pyqt5 python3-dev ' \
              'python3-distutils-extra gir1.2-gexiv2-0.10 python3-gi gir1.2-gudev-1.0 ' \
              'gir1.2-udisks-2.0 gir1.2-notify-0.7 gir1.2-glib-2.0 gir1.2-gstreamer-1.0 '\
              'libgphoto2-dev python3-arrow python3-psutil g++ libmediainfo0v5 '\
@@ -254,12 +270,16 @@ def check_package_import_requirements(distro: Distro, version: float) -> None:
         if missing_packages:
             cmd = shutil.which('apt-get')
             command_line = 'sudo {} install {}'.format(cmd, ' '.join(missing_packages))
-            install_packages(command_line)
+            print("To continue, some packages required to run the application will be "
+                  "installed.\n")
+            run_cmd(command_line)
 
-    elif distro in (Distro.fedora,):
+    elif distro in fedora_like:
         if not have_dnf:
-            sys.stderr.write('To continue, the package python3-dnf must be installed.')
-            sys.exit(1)
+            print('To continue, the package python3-dnf must be installed.\n')
+            cmd = shutil.which('dnf')
+            command_line = 'sudo {} install python3-dnf'.format(cmd)
+            run_cmd(command_line, restart=True)
 
         missing_packages = []
         packages = 'python3-qt5 gobject-introspection python3-gobject ' \
@@ -268,7 +288,7 @@ def check_package_import_requirements(distro: Distro, version: float) -> None:
                    'python3-easygui qt5-qtimageformats python3-psutil libmediainfo ' \
                    'python3-requests'.split()
 
-        if version <= 24.0 and version > 0.0:
+        if 0.0 < version <= 24.0:
             packages.append('libgexiv2-python3')
         else:
             packages.append('python3-gexiv2')
@@ -307,12 +327,30 @@ def check_package_import_requirements(distro: Distro, version: float) -> None:
         if missing_packages:
             cmd = shutil.which('dnf')
             command_line = 'sudo {} install {}'.format(cmd, ' '.join(missing_packages))
-            install_packages(command_line)
+            print("To continue, some packages required to run the application will be "
+                  "installed.\n")
+            run_cmd(command_line)
+
+    elif distro == Distro.opensuse:
+        cmd = shutil.which('zypper')
+        packages = 'python3-qt5 girepository-1_0 python3-gobject ' \
+                   'zeromq-devel exiv2 exiftool python3-devel python3-distutils-extra ' \
+                   'libgphoto2-devel libraw-devel gcc-c++ rpm-build intltool ' \
+                   'libqt5-qtimageformats python3-requests python3-psutil ' \
+                   'typelib-1_0-GExiv2-0_10 typelib-1_0-UDisks-2_0 typelib-1_0-Notify-0_7 ' \
+                   'typelib-1_0-Gst-1_0 typelib-1_0-GUdev-1_0  '
+        command_line = 'sudo {} in {}'.format(cmd, packages)
+        print("To continue, some packages required to run the application will be checked or "
+              "installed.\n")
+
+        run_cmd(command_line)
+
+        #TODO libmediainfo - not a default openSUSE package, sadly
     else:
         check_packages_on_other_systems()
 
 def query_uninstall() -> bool:
-    return input('Type yes and hit enter if you want to to uninstall the previous version of '
+    return input('\nType yes and hit enter if you want to to uninstall the previous version of '
                  'Rapid Photo Downloader: ') == 'yes'
 
 def uninstall_old_version(distro: Distro) -> None:
@@ -324,9 +362,9 @@ def uninstall_old_version(distro: Distro) -> None:
         if pkg.is_installed and query_uninstall():
             cmd = shutil.which('apt-get')
             command_line = 'sudo {} remove {}'.format(cmd, pkg_name)
-            uninstall_packages(command_line)
+            run_cmd(command_line)
 
-    elif distro == Distro.fedora:
+    elif distro in fedora_like:
         print("Querying package system to see if an older version of Rapid Photo Downloader is "
               "installed (this may take a while)...")
         with dnf.Base() as base:
@@ -338,7 +376,18 @@ def uninstall_old_version(distro: Distro) -> None:
             if len(list(i)) and query_uninstall():
                 cmd = shutil.which('dnf')
                 command_line = 'sudo {} remove {}'.format(cmd, pkg_name)
-                uninstall_packages(command_line)
+                run_cmd(command_line)
+
+    elif distro == Distro.opensuse:
+        print("Querying package system to see if an older version of Rapid Photo Downloader is "
+              "installed (this may take a while)...")
+        zypper = shutil.which('zypper')
+        command_line = '{} se rapid-photo-downloader'.format(zypper)
+        args = shlex.split(command_line)
+        output = subprocess.check_output(args, universal_newlines=True)
+        if '\ni | rapid-photo-downloader' in output and query_uninstall():
+            command_line = 'sudo {} rm rapid-photo-downloader'.format(zypper)
+            run_cmd(command_line)
 
 
 def make_pip_command(args: str):
@@ -391,7 +440,7 @@ def main(installer: str, distro: Distro, distro_version: float) -> None:
     install_path = os.path.join(os.path.expanduser('~'), '.local', 'bin')
 
     if install_path not in path.split(':'):
-        if distro in debian_like:
+        if distro in debian_like or distro == Distro.opensuse:
             bin_dir = os.path.join(os.path.expanduser('~'), 'bin')
             if not os.path.isdir(bin_dir):
                 created_bin_dir = True
@@ -456,36 +505,71 @@ if __name__ == '__main__':
     else:
         distro_version = 0.0
 
-    if distro == Distro.debian and distro_version == 8.0:
+    if distro == Distro.debian and distro_version <= 8.0:
         sys.stderr.write("Sorry, Debian Jessie is too old to be able to run this version of Rapid "
                          "Photo Downloader.\n")
         sys.exit(1)
-    elif distro == Distro.fedora and distro_version <= 23.0:
+    elif distro in fedora_like and 0.0 > distro_version <= 23.0:
         sys.stderr.write("Sorry, Fedora 23 is no longer supported by Rapid Photo Downloader.\n")
         sys.exit(1)
+    elif distro in arch_like:
+        print('Users of Arch Linux or its derivatives should try the AUR package: '
+              'https://aur.archlinux.org/packages/rapid-photo-downloader-bzr/')
+        print("Exiting...")
+        sys.exit(0)
+
+    if distro in debian_like:
+        distro_family = Distro.debian
+    elif distro in fedora_like:
+        distro_family = Distro.fedora
+    else:
+        distro_family = distro
+
+    packages = []
+    try:
+        import pip
+    except ImportError:
+        packages.append('python3-pip')
+
+    try:
+        import setuptools
+    except ImportError:
+        packages.append('python3-setuptools')
+
+    try:
+        import wheel
+    except:
+        packages.append('python3-wheel')
+
+    if packages:
+        packages = ' '.join(packages)
+
+        if distro_family not in (Distro.fedora, Distro.debian, Distro.opensuse):
+            sys.stderr.write(
+                "Install the following packacges using your Linux distribution's standard package "
+                "manager, and then rerun this installer\n")
+            sys.stderr.write(packages + '\n')
+            sys.exit(1)
+
+        print("To run this program, you must first install some programs to assist "
+              "Python 3 and its package management.\n")
+
+        installer = installer_cmds[distro_family]
+        command_line = 'sudo {} install '.format(installer) + packages
+        run_cmd(command_line, restart=True)
+
+    # Can now assume that both pip and wheel have been installed
 
     if StrictVersion(pip.__version__) < StrictVersion('8.1'):
-        sys.stderr.write("Python 3's pip and setuptools must be upgraded for your user.\n\n")
+        print("\nPython 3's pip and setuptools must be upgraded for your user.\n")
 
-        if get_distro() == Distro.fedora:
-            sys.stderr.write("\nNOTE: ensure the package python3-wheel is installed before you "
-                             "upgrade pip and setuptools. Install it by running this command:\n"
-                             "sudo dnf install python3-wheel\n\n")
+        print("Caution: upgrading pip and setuptools for your user could potentially "
+             "negatively affect the installation of other, older Python packages by your user.\n")
+        print("However the risk is very small and is normally nothing to worry about.\n")
 
-        sys.stderr.write('To upgrade pip and setuptools, run both of these commands as your '
-                         'regular user (i.e. without sudo):\n\n')
-           
-        for cmd in ('python3 -m pip install --user --upgrade pip',
-                    'python3 -m pip install --user --upgrade setuptools'):
-            sys.stderr.write('{}\n'.format(cmd))
-        sys.stderr.write('\nThen run this installer again.\n\n')
+        command_line = '{} -m pip install --user --upgrade pip setuptools'.format(sys.executable)
 
-        sys.stderr.write("Caution: upgrading pip and setuptools for your user could potentially "
-                         "negatively affect the installation of other, older Python packages by your user.\n")
-        sys.stderr.write("However the risk is small and is normally nothing to worry about.\n")
-
-
-        sys.exit(1)
+        run_cmd(command_line, restart=True)
 
     parser = argparse.ArgumentParser(description='Install Rapid Photo Downloader')
     parser.add_argument('tarfile', action='store', help="tar.gz Rapid Photo Downloader "
@@ -494,6 +578,7 @@ if __name__ == '__main__':
     installer = args.tarfile  # type: str
     if not os.path.exists(installer):
         print("Installer not found:", installer)
+        print("Include the name of the tar.gz Rapid Photo Downloader installer archive")
         sys.exit(1)
     elif not installer.endswith('.tar.gz'):
         print("Installer not in tar.gz format:", installer)